api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysqlclient==2.2.0
redis==5.0.1
python-dotenv==1.0.0
